2025-05-22-001
Changed the naming to use serialno instead of computer name as computer name isn't relevant to autopilot at all.
Added logic to install the script, still need to make logic to look for it in local path and then just use it rather than install and install prompts for nuget add in OOBE, that's a bit annoying.